streamlit_sql_sandbox/requirements.txt
altair
llama-index-core>=0.11.0
# see https://llamahub.ai for all integrations
llama-index-embeddings-openai  # can replace with the embed model you want to use
llama-index-llms-openai  # can replace with the llm you want to use
llama-index-agent-openai
streamlit>=1.37.1
